Finished Project (final commit)
On the index page, I have created the entire "home" page and inserted the necessary information and links such as information on the company, it's address, links to email/call them, et cetera. On the jtc_services_html page I have created the "services" page which was a list of the certain things that the company is able to do.